refactor: ui 모듈에 ui 접두사 추가
