+++ feature-ui-notification/src/main/AndroidManifest.xml
