+++ feature-ui-onboard/src/main/AndroidManifest.xml
