+++ feature-ui-setting/src/main/AndroidManifest.xml
